requirements.txt
python-dotenv
pygame
gymnasium
numpy
matplotlib